requirements.txt
# Core app libraries
streamlit==1.35.0
pandas==2.2.2
numpy==1.26.4

# Sentence transformer and FAISS for semantic search
sentence-transformers==2.6.1
faiss-cpu==1.7.4  # If you're using a GPU, replace with faiss-gpu

# Google Gemini API
google-generativeai==0.5.2

# Environment and utilities
python-dotenv==1.0.1
requests==2.31.0
beautifulsoup4==4.12.3
